==> code/01_CourseScraper.ipynb <==
# finds which courses should be scraped

import pandas as pd
import requests
from bs4 import BeautifulSoup

# gets department names
URL = 'https://www.layuplist.com/departments'

page = requests.get(URL)
soup = BeautifulSoup(page.content, "html.parser")
html = requests.get(URL).content
html_table = pd.read_html(html)
course_code_db = html_table[-1]
course_code_db

# list to be converted into df
course_db_list = []

# goes through each department
for depart in course_code_db['Code']:
    URL = 'https://www.layuplist.com/search?q=' + depart
    html = requests.get(URL).content
    
    # reads table of courses in department
    df_list = pd.read_html(html)
    df2 = df_list[-1]
    course_db_list.append(df2)
    print('Finished scraping ', depart)

# combines all courses from all departments into dataframe
course_db = pd.concat(course_db_list)

course_db

# cleans dataset
course_db = course_db[course_db['Reviews'] > 0] 
course_db = course_db.reset_index()
course_db = course_db.drop(columns=['index', 0,'Layup', 'Offered 22S?'])


course_db.to_csv ('../data/course_db.csv', index = False, header=True)
course_code_db.to_csv ('../data/course_code_db.csv', index = False, header=True)
