==> code/02.01_LayupListScraper.ipynb <==
import pandas as pd
import requests
from bs4 import BeautifulSoup
import html5lib
import time
from bs4 import BeautifulSoup
from selenium import webdriver
from dotenv import load_dotenv

# uses env file to store LayupList login information
load_dotenv()

LAYUPLIST_EMAIL = os.getenv('LAYUPLIST_EMAIL')
LAYUPLIST_PASSWORD = os.getenv('LAYUPLIST_PASSWORD')

# reads in course and department catalog
course_db = pd.read_csv ('../data/course_db.csv')
course_code_db = pd.read_csv ('../data/course_code_db.csv')

# logs into layuplist
driver = webdriver.Chrome('/Users/sarakay/Downloads/chromedriver')  # Optional argument, if not specified will search path.
driver.get('https://www.layuplist.com/accounts/login/');


driver.find_element_by_id("login-email").send_keys(LAYUPLIST_EMAIL)

driver.find_element_by_id("login-password").send_keys(LAYUPLIST_PASSWORD)

driver.find_element_by_xpath("/html/body/div/div[1]/div/div/div/div[5]/div/form/input[4]").click() 

import re

driver.get('https://www.layuplist.com/departments')

# goes through each department name
for depart in course_code_db['Code']: 
    
    # does to page on layuplist containing department's courses
    URL = 'https://www.layuplist.com/search?q=' + depart
    driver.get(URL)
    
    # gets courses in department
    course_names = course_db[course_db['Course'].str.startswith(depart)]['Course']
    
    # goes to each course 
    for departCourse in course_names:
        driver.get(URL)
        print(departCourse)
        
        try:
            driver.find_element_by_partial_link_text(departCourse).click()
            soup = BeautifulSoup(driver.page_source)
            
            # gets number of pages of reviews to scrape
            maxNum = 1
            numRange = soup.findAll('ul')
            num = []
            urlStarter = driver.current_url
            for numbers in numRange:
                if 'of' in numbers.text:
                    num = re.findall('[0-9]+', numbers.text)
                    maxNum = int(num[1])
                    
            # list to store each page of reviews
            listToDF = []
            for i in range(1, maxNum + 1):
                newURL = urlStarter + '?page=' + str(i)
                driver.get(newURL)
                soup = BeautifulSoup(driver.page_source)

                paragraphs = soup.findAll('td')
                
                # puts review text into list
                for paragraph in paragraphs:
                    if paragraph.b != None:
                        listToDF.append(paragraph.text)
                        
            # creates dataframe from list of reviews
            coursedf = pd.DataFrame(listToDF, columns=['courseReviewRaw'])
            
            # labels which course review came from
            coursedf['Course'] = departCourse
            complete_df_list.append(coursedf)
        except:
            continue

complete_df_list

layuplistreviews = pd.concat(complete_df_list)
layuplistreviews

layuplistreviews.to_csv ('../data/layuplist_reviews.csv', index = False, header=True)
