==> code/03.01_LayupListDataCleaner.ipynb <==
# reads in layuplist coures reviews
import pandas as pd
layuplistdb = pd.read_csv('../data/layuplist_reviews.csv')
layuplistdb

# sample raw test from layuplist
test = layuplistdb.courseReviewRaw[5]
test.split('\n')

# gets department name from Course column
def extract_department(row):
    s = row['Course'].split(':')[0]
    result = ''.join([i for i in s if not i.isdigit()])
    return result

# gets term based on placement in review text
def extract_term(row):
    return row['courseReviewRaw'].split('\n')[1].strip()

# gets professor name based on placement and keyword 'with'
# catches error where course does not have professor 
def extract_prof(row):
    try:
        return row['courseReviewRaw'].split('\n')[2].split("with",1)[1].strip()[:-1]
    except:
        return None

# gets review from all sections after professor
def extract_review(row):
    text_list = row['courseReviewRaw'].split('\n')[3:]
    return ''.join(text_list).strip()

layuplistdb['Department'] = layuplistdb.apply(extract_department, axis = 1)
layuplistdb['Term'] = layuplistdb.apply(extract_term, axis = 1)
layuplistdb['Prof'] = layuplistdb.apply(extract_prof, axis = 1)
layuplistdb['Review'] = layuplistdb.apply(extract_review, axis = 1)

layuplistdb = layuplistdb.drop(columns=['courseReviewRaw'])

layuplistdb = layuplistdb.dropna()

layuplistdb

layuplistdb.to_csv('../data/layuplist_cleaned.csv')
