==> code/03.02_CourseAssessmentDataCleaner.ipynb <==
# cleans up courseassessmentreviews
import pandas as pd
courseassessmentdb = pd.read_csv('../data/course_assessment_reviews.csv')
courseassessmentdb

# reformats the term to match with LayupList term format
def clean_term(row):
    term_info = row['Term'].split(' ')
    
    # changes ending from S to X based on summer term
    if term_info[0] != 'Summer':
        return term_info[2][2:] + term_info[0][0]
    else:
        return term_info[2][2:] + 'X'
    
courseassessmentdb['Term'] = courseassessmentdb.apply(clean_term, axis = 1)
courseassessmentdb

courseassessmentdb.to_csv('../data/course_assessment_cleaned.csv')
