==> code/03.03_CourseReviewMerger.ipynb <==
# merges the two datasets

import pandas as pd
layuplistdb = pd.read_csv('../data/layuplist_cleaned.csv', index_col = 0)
layuplistdb

courseassessmentdb = pd.read_csv('../data/course_assessment_cleaned.csv', index_col = 0)
courseassessmentdb

# uses concat to perform the merge
combiner = [layuplistdb, courseassessmentdb]
combined = pd.concat(combiner)
combined = combined.dropna()
combined

combined.to_csv('../data/course_review_db.csv')
