==> code/03.04 Professor&CourseFuzzyNameMatching.ipynb <==
# scrapes faculty directory for prof names

import requests
from bs4 import BeautifulSoup
import pandas as pd


faculty_name = []

# manually got the max num of pages in directory
for i in range(1, 34):
    URL = 'https://faculty-directory.dartmouth.edu/department?dept=All&page=' + str(i)

    page = requests.get(URL)
    soup = BeautifulSoup(page.content, "html.parser")

    result = soup.find_all('span', 'class' == 'field-content')
    faculty_name += [r.text for r in result if r.text != '']

# df of all faculty at Dartmouth
faculty_name = set(faculty_name)
prof_df = pd.DataFrame(faculty_name, columns=['ProfessorName'])
prof_df

# reads in course reviews for professor names
import pandas as pd
course_review_db = pd.read_csv('../data/course_review_db.csv', index_col = 0)
course_review_db

# Sample professor who has multiple names
course_review_db[course_review_db['Prof'] == 'Sebastiaan Joosten']

course_review_db[course_review_db['Prof'] == 'Sebastiaan Joze Joosten']

from fuzzywuzzy import fuzz
import pandas as pd
import numpy as np

# sets up linking column between directory and review dataframe
dataframecolumn = pd.DataFrame(prof_df['ProfessorName'])
dataframecolumn.columns = ['directory_name']
compare = pd.DataFrame(set(course_review_db['Prof']))
compare.columns = ['fuzzy_match_name']

# merges names into combined dataframe
dataframecolumn['Key'] = 1
compare['Key'] = 1
combined_dataframe = dataframecolumn.merge(compare,on="Key",how="left")
combined_dataframe = combined_dataframe[~(combined_dataframe.directory_name==combined_dataframe.fuzzy_match_name)]

def partial_match(x,y):
    return(fuzz.ratio(x,y))
partial_match_vector = np.vectorize(partial_match)

# looks for names that are likely to match on directory name, ignore exact match
combined_dataframe['score']=partial_match_vector(combined_dataframe['directory_name'],combined_dataframe['fuzzy_match_name'])
combined_dataframe = combined_dataframe[combined_dataframe.score>=85]
combined_dataframe = combined_dataframe[combined_dataframe.score!=100]
combined_dataframe

# sample name that fuzzy matches to two different names with same directory name
combined_dataframe[combined_dataframe['directory_name'] == 'Sebastiaan J. C. Joosten'].sort_values('score')

# adjusts professor name based on fuzzy matching results
def fuzzmatcher(row):
    if combined_dataframe['fuzzy_match_name'].str.startswith(row['Prof']).any():
        return list(combined_dataframe[combined_dataframe['fuzzy_match_name'].str.startswith(row['Prof'])]['directory_name'])[0]
    else:
        return row['Prof']
    
course_review_db['Prof'] = course_review_db.apply(fuzzmatcher, axis = 1)
course_review_db

# newly merged professor name
course_review_db[course_review_db['Prof'] == 'Sebastiaan J. C. Joosten']

prof_df.to_csv('../data/professor_directory.csv')

# Example of mismatch course number/name
course_review_db[course_review_db['Course'].str.contains('Oceania')]

course_db = pd.read_csv('../data/course_db.csv')
course_db

from fuzzywuzzy import fuzz
import pandas as pd
import numpy as np

# sets up linking column between directory and review dataframe
dataframecolumn = pd.DataFrame(course_db['Course'])
dataframecolumn.columns = ['directory_name']
compare = pd.DataFrame(set(course_review_db['Course']))
compare.columns = ['fuzzy_match_name']

# merges names into combined dataframe
dataframecolumn['Key'] = 1
compare['Key'] = 1
combined_dataframe = dataframecolumn.merge(compare,on="Key",how="left")
combined_dataframe = combined_dataframe[~(combined_dataframe.directory_name==combined_dataframe.fuzzy_match_name)]

def partial_match(x,y):
    return(fuzz.ratio(x,y))
partial_match_vector = np.vectorize(partial_match)

# looks for names that are likely to match on directory name, ignore exact match
combined_dataframe['score']=partial_match_vector(combined_dataframe['directory_name'],combined_dataframe['fuzzy_match_name'])
combined_dataframe = combined_dataframe[combined_dataframe.score>=85]
combined_dataframe = combined_dataframe[combined_dataframe.score!=100]
combined_dataframe

# adjusts professor name based on fuzzy matching results
def fuzzmatcher2(row):
    if combined_dataframe['fuzzy_match_name'].str.startswith(row['Course']).any():
        return list(combined_dataframe[combined_dataframe['fuzzy_match_name'].str.startswith(row['Course'])]['directory_name'])[0]
    else:
        return row['Course']
    
course_review_db['Course'] = course_review_db.apply(fuzzmatcher2, axis = 1)
course_review_db

course_review_db.to_csv('../data/course_review_db_prof_course_name_cleaned.csv')
