==> code/03.05 ProfessorGenderClassification.ipynb <==
# reads in course reviews for professor names
import pandas as pd
course_review_db = pd.read_csv('../data/course_review_db_prof_course_name_cleaned.csv', index_col = 0)
course_review_db

import re

def words_in_string(word_list, a_string):
    return len(set(word_list).intersection(a_string.split())) > 0

# assigns professor gender based on the pronouns used within review
def determine_prof_gender(row):
    raw_text = row['Review'].lower()
    
    male_words = ['he', 'his', 'himself']
    female_words = ['she', 'her', 'herself']
    
    if words_in_string(male_words, raw_text):
        return 'Male'
    elif words_in_string(female_words, raw_text):
        return 'Female'
    else:
        return None
    
course_review_db['Prof Gender'] = course_review_db.apply(determine_prof_gender, axis = 1)
course_review_db

# counts how many times a professor was tagged as for each gender 
prof_directory = course_review_db.groupby(['Prof', 'Prof Gender']).count().reset_index()
prof_directory = prof_directory.groupby('Prof', group_keys=False).apply(lambda x: x.loc[x['Course'].idxmax()])
prof_directory = prof_directory.drop(columns=['Prof']).reset_index()
prof_directory

# sample professor that was classified as both female and male, but majority female
prof_directory = course_review_db.groupby(['Prof', 'Prof Gender']).count().reset_index()
prof_directory[prof_directory.Prof == 'Lynda Boose']

# reassigns professors with gender count that is the highest
course_review_db = pd.merge(course_review_db,prof_directory[['Prof','Prof Gender']],on='Prof', how='left')
course_review_db = course_review_db.drop(columns=['Prof Gender_x'])
course_review_db = course_review_db.rename(columns={"Prof Gender_y": "Prof Gender"})
course_review_db

course_review_db[(course_review_db['Prof Gender'] != 'Male') & (course_review_db['Prof Gender'] != 'Female')]

import random
from nltk.corpus import names
import nltk

def gender_features(word):
    return {'last_letter':word[-1]}

# adds already tagged male professors to data set
males = set([x.split(' ')[0] for x in course_review_db[course_review_db['Prof Gender'] == 'Male']['Prof']] )
males = [m for m in males if len(m) > 1 ]

# adds already tagged female professors to data set
females = set([x.split(' ')[0] for x in course_review_db[course_review_db['Prof Gender'] == 'Female']['Prof']] )
females = [f for f in females if len(f) > 1 ]

# uses names from nltk corpus to create classifier for gender 
labeled_names = ([(name, 'male') for name in names.words('male.txt')] + [(name, 'female') for name in names.words('female.txt')])
Dartmouth_labeled_names =  ([(name, 'male') for name in males] + [(name, 'female') for name in females])
labeled_names = labeled_names + Dartmouth_labeled_names
random.shuffle(labeled_names)

# creates training and testing set for names 
featuresets = [(gender_features(n), gender) for (n, gender) in labeled_names]
train_set, test_set = featuresets[500:], featuresets[:500]
classifier = nltk.NaiveBayesClassifier.train(train_set)

# uses Naive Bayes model to clasify professor gender

def determine_prof_gender2(row):
    try:
        prof_first_name = row['Prof'].split(' ')[0]
        return classifier.classify(gender_features(prof_first_name)).capitalize()
    except:
        return None
    
course_review_db['Prof Gender'] = course_review_db.apply(lambda row: determine_prof_gender2(row) if (row['Prof Gender'] != 'Male') & (row['Prof Gender'] != 'Female') else row['Prof Gender'], axis=1)
course_review_db

course_review_db[(course_review_db['Prof Gender'] != 'Male') & (course_review_db['Prof Gender'] != 'Female')]

course_review_db.to_csv('../data/course_db_cleaned_final.csv')
